requirement.txt
django==4.2.1
djangorestframework==3.14.0
requests-2.30.0